Rework flavor resolution
In all cases, we use an absolute path to the pytest executable used to
run the test suite. This path uses the known location for the `bundle`
flavor. The `python3` flavor is resolved to a version flavor like
python313 when the "primary" Python interpreter is not called "python3".

pytest is invoked via /usr/bin/python3 -m pytest when the flavor is
python3. The reason is that, on SLE 15, default pytest selected by
update-alternatives might be /usr/bin/pytest -> /usr/bin/pytest-3.11.

This unifies the approach for both `-f bundle` and `-f python3*`,
without having to update $PATH in the former case.

flavor=python3 -> '/usr/bin/python3 -m pytest'